seedData/requirements.txt
faker
mysql.connector
